+++ app/Actions/Fortify/ResetUserPassword.php
@@ -2,15 +2,13 @@
 
 namespace App\Actions\Fortify;
 
-use App\Concerns\PasswordValidationRules;
+use App\Http\Requests\Auth\ResetUserPasswordRequest;
 use App\Models\User;
 use Illuminate\Support\Facades\Validator;
 use Laravel\Fortify\Contracts\ResetsUserPasswords;
 
 class ResetUserPassword implements ResetsUserPasswords
 {
-    use PasswordValidationRules;
-
     /**
      * Validate and reset the user's forgotten password.
      *
@@ -18,9 +16,12 @@ class ResetUserPassword implements ResetsUserPasswords
      */
     public function reset(User $user, array $input): void
     {
-        Validator::make($input, [
-            'password' => $this->passwordRules(),
-        ])->validate();
+        Validator::make(
+            $input,
+            ResetUserPasswordRequest::baseRules(),
+            [],
+            ResetUserPasswordRequest::attributeNames(),
+        )->validate();
 
         $user->forceFill([
             'password' => $input['password'],