+++ app/Http/Requests/Auth/ResetUserPasswordRequest.php
@@ -1,0 +1,56 @@
+<?php
+
+namespace App\Http\Requests\Auth;
+
+use Illuminate\Contracts\Validation\ValidationRule;
+use Illuminate\Foundation\Http\FormRequest;
+use Illuminate\Validation\Rules\Password;
+
+class ResetUserPasswordRequest extends FormRequest
+{
+    /**
+     * Determine if the user is authorized to make this request.
+     */
+    public function authorize(): bool
+    {
+        return true;
+    }
+
+    /**
+     * Get the validation rules that apply to the request.
+     *
+     * @return array<string, ValidationRule|array<mixed>|string>
+     */
+    public function rules(): array
+    {
+        return self::baseRules();
+    }
+
+    /**
+     * @return array<string, ValidationRule|array<mixed>|string>
+     */
+    public static function baseRules(): array
+    {
+        return [
+            'password' => ['required', 'string', Password::default(), 'confirmed'],
+        ];
+    }
+
+    /**
+     * @return array<string, string>
+     */
+    public function attributes(): array
+    {
+        return self::attributeNames();
+    }
+
+    /**
+     * @return array<string, string>
+     */
+    public static function attributeNames(): array
+    {
+        return [
+            'password' => __('auth.fields.password'),
+        ];
+    }
+}